requirements.txt
#requirements.txt

#tools
cython==0.19.2
edn_format==0.4
pexpect==2.4
docopt==0.6.1
termcolor==1.1.0
docutils==0.11
numpy==1.7.1
sympy==0.7.3
pil==1.1.7
python-dateutil==2.1
pytz==2013.7
xhtml2pdf==0.0.5
boto==2.17.0
Unidecode==0.04.14
selenium==2.37.2
Scrapy==0.18.4

#web
requests==2.0.0
gunicorn==18.0
#py-postgresql==1.1.0
psycopg2==2.5.1
psycogreen==1.0
sqlalchemy==0.8.2
zope.sqlalchemy==0.7.2
transaction==1.4.1
alembic==0.6.0
futures==2.1.5
tornado==3.1.1
gevent==0.13.8
gevent_subprocess==0.1.1
gevent-socketio==0.3.5-rc2
pyramid_socketio==0.9
pyramid==1.4.3
pyramid_tm==0.7
dogpile.cache==0.5.1
pylibmc==1.2.3
pyramid_beaker==0.8
MemcachedManager==0.2
pyramid_redis_sessions==0.9b2
pyramid_debugtoolbar==1.0.6
pyramid_exclog==0.7
jinja2==2.7.1
pyramid_jinja2==1.8
pyramid_simpleform==0.6.1
pyramid_deform==0.2a5
pyramid_mailer==0.13
passlib==1.6.1
velruse==1.1
pyramid_persona==1.5
waitress==0.8.5
paste==1.7.5.1
pyzmq==14.0.0
u-msgpack-python==1.4
#uwsgi==1.9.18.2
#m2wsgid==0.7.1
elasticsearch==0.4.3
pyelasticsearch==0.6.1
sunburnt==0.6
envoy==0.0.2

#test
nose==1.3.0
WebTest==2.0.6
coverage==3.6
behave==1.2.3
mock==1.0.1

#static content
webassets==0.9
pyScss==1.2.0.post3
pyramid_webassets==0.7.1
closure==20121212
yuicompressor==2.4.8
jsmin==2.0.8
slimmer==0.1.30
cssutils==0.9.10
cssmin==0.2.0

#background
pymongo==2.6.3
pyzmq-mdp==0.2
#mongokit==0.8.3
#redis==2.8.0
#pyres==1.4.2
#celery==3.0.21
#celery-with-redis==3.0
#flower==0.5.2
#pytz==2013b

-e ./oasysusa

